Trained a new model + minor bug fix
Trained a new model that can successfully enter the rotating corridor. Saved as 'rdv_model_rot.zip'. This model uses a different reward function (see logs\notes.md for more info).

Fixed a minor bug in custom_env that was affecting the training on Colab (probably because they updated numpy).

Modified plot_learning to show a red marker on the highest reward achieved.

+++ custom_env.py
@@ -45,10 +45,10 @@ class Rendezvous3DOF(gym.Env):
         self.t_max = 300        # Max time per episode
 
         # Orbit properties:
-        self.mu = 3.986004418e14    # Gravitational parameter of Earth (m3/s2)
-        self.Re = 6371e3            # Radius of the Earth (m)
-        self.h = 800e3              # Altitude of the orbit (m)
-        self.r = self.Re + self.h   # Radius of the orbit (m)
+        self.mu = 3.986004418e14            # Gravitational parameter of Earth (m3/s2)
+        self.Re = 6371e3                    # Radius of the Earth (m)
+        self.h = 800e3                      # Altitude of the orbit (m)
+        self.r = self.Re + self.h           # Radius of the orbit (m)
         self.n = sqrt(self.mu / self.r**3)  # Mean motion of the orbit (rad/s)
 
         # Action and state constraints:
@@ -132,7 +132,7 @@ class Rendezvous3DOF(gym.Env):
 
         # Observation: the observation is equal to the normalized state
         # obs = self.state
-        obs = self.normalize_state()
+        obs = self.normalize_state().astype(self.observation_space.dtype)
 
         # Reward: for now, the reward penalizes distance from the origin
         dist = self.absolute_position()
